+++ cmd/squaringd/commands/start.go
@@ -7,13 +7,11 @@ import (
 	"path/filepath"
 
 	"github.com/0xPellNetwork/pellapp-sdk/client"
-	clienthelpers "github.com/0xPellNetwork/pellapp-sdk/client/helpers"
 	"github.com/0xPellNetwork/pellapp-sdk/server"
 	servertypes "github.com/0xPellNetwork/pellapp-sdk/server/types"
 	"github.com/0xPellNetwork/pelldvs-libs/log"
 	pelldvscfg "github.com/0xPellNetwork/pelldvs/config"
 	dbm "github.com/cosmos/cosmos-db"
-	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"golang.org/x/sync/errgroup"
@@ -24,20 +22,12 @@ import (
 	sqserver "github.com/0xPellNetwork/dvs-template/dvs/squared/server"
 )
 
-var DefaultNodeHome string
-
-func init() {
-	var err error
-	DefaultNodeHome, err = clienthelpers.GetNodeHomeDirectory(".pelldvs")
-	if err != nil {
-		panic(err)
-	}
-}
-
 func addFlags(cmd *cobra.Command) {
 	cmd.Flags().String("squared-config", "", "Path to the squared config file")
 }
 
+// execPostSetup initializes and configures various components of the server after setup, such as loading configurations,
+// creating chain connector clients, and starting the task dispatcher. It runs concurrent tasks using an error group.
 func execPostSetup(svrCtx *server.Context, clientCtx client.Context, ctx context.Context, app servertypes.Application, g *errgroup.Group) error {
 	var configPath = DefaultNodeHome + "/config/config.toml"
 	var pellDVSConfig pelldvscfg.Config
@@ -87,6 +77,7 @@ func execPostSetup(svrCtx *server.Context, clientCtx client.Context, ctx context
 	return nil
 }
 
+// postSetup is called after the application is initialized
 func postSetup(svrCtx *server.Context, clientCtx client.Context, ctx context.Context, app servertypes.Application, g *errgroup.Group) error {
 	// start dispatcher here
 	g.Go(func() error {
@@ -95,16 +86,16 @@ func postSetup(svrCtx *server.Context, clientCtx client.Context, ctx context.Con
 	return nil
 }
 
+// openDB opens the database
 func openDB(rootDir string, backendType dbm.BackendType) (dbm.DB, error) {
 	dataDir := filepath.Join(rootDir, "data")
 	return dbm.NewDB("squaredapp", backendType, dataDir)
 }
 
+// InitRunOperatorCommand initializes the run operator command
 func InitRunOperatorCommand(rootCmd *cobra.Command) {
-	cmd := runOperatorCmdV2(rootCmd)
-
 	server.AddCommandsWithStartCmdOptions(
-		cmd,
+		rootCmd,
 		DefaultNodeHome,
 		newApp,
 		server.StartCmdOptions{
@@ -114,8 +105,6 @@ func InitRunOperatorCommand(rootCmd *cobra.Command) {
 			StartCommandHandler: nil,
 		},
 	)
-
-	rootCmd.AddCommand(cmd)
 }
 
 // newApp creates the application
@@ -133,38 +122,4 @@ func newApp(
 		appOpts,
 		baseappOptions...,
 	)
-}
-
-func runOperatorCmdV2(rootCmd *cobra.Command) *cobra.Command {
-	initClientCtx := client.Context{}.
-		WithHomeDir(DefaultNodeHome).
-		WithInterfaceRegistry(codectypes.NewInterfaceRegistry())
-
-	cmd := &cobra.Command{
-		Use:     "run-operator",
-		Aliases: []string{"r"},
-		Short:   "Run the operator",
-		Long:    "Run the operator service",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return nil
-		},
-		PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
-			// set the default command outputs
-			cmd.SetOut(cmd.OutOrStdout())
-			cmd.SetErr(cmd.ErrOrStderr())
-
-			initClientCtx = initClientCtx.WithCmdContext(cmd.Context())
-
-			if err := client.SetCmdClientContextHandler(initClientCtx, cmd); err != nil {
-				return err
-			}
-
-			customAppTemplate, customAppConfig := initAppConfig()
-			customPellDVSConfig := initPellDVSConfig()
-
-			return server.InterceptConfigsPreRunHandler(cmd, customAppTemplate, customAppConfig, customPellDVSConfig)
-		},
-	}
-
-	return cmd
 }
